Work in progress on the meta-bebot layer
This commit adds enough support for the ISS and the OV5640 to detect
each other and start communicating. The patch set for uboot and linux
has been cleaned up to remove the patches for the OMAP3 target.

Streaming crashes at the moment due to a memory / DMA / CMA issue,
however I2C communication with the camera is verified and we have
/dev/videoX nodes and working media-ctl -p